+++ src/util/figure.test.ts
@@ -60,11 +60,11 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-b-0.0.0.0" data-group="fig" data-index="1.0.0.1" id="label:fig-1.0.0.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.0.0.1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-c-0.0.0" data-group="fig" data-index="1.0.1" id="label:fig-1.0.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.0.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-b-0.0.0.0" data-group="fig" data-index="1.1.1.1" id="label:fig-1.1.1.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-c-0.0.0" data-group="fig" data-index="1.1.2" id="label:fig-1.1.2"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.2.</span><figcaption></figcaption></figure>',
             '<figure data-label="fig-d" data-group="fig" data-index="2" id="label:fig-2"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 2.</span><figcaption></figcaption></figure>',
-            '<p><a href="#label:fig-1.0.0.1" rel="noopener" class="label" data-label="fig-b">Fig. 1.0.0.1</a></p>',
+            '<p><a href="#label:fig-1.1.1.1" rel="noopener" class="label" data-label="fig-b">Fig. 1.1.1.1</a></p>',
           ]);
       }
     });
@@ -73,9 +73,9 @@ describe('Unit: util/figure', () => {
       const source = parse([
         '~~~figure [:fig-1]\n!https://host\n~~~',
         '~~~figure [:fig-a-0.0]\n!https://host\n~~~',
-        '~~~figure [:$-2.0]\n$$\nLaTeX\n$$\n~~~',
-        '~~~figure [:fig-b-0.0]\n!https://host\n~~~',
+        '~~~figure [:$-2.0.0]\n$$\nLaTeX\n$$\n~~~',
         '~~~figure [:quote-a-0.0]\n> \n~~~',
+        '~~~figure [:fig-b-0.0.0]\n!https://host\n~~~',
         '~~~figure [:$-3.1]\n$$\nLaTeX\n$$\n~~~',
         '[:fig-2.0]',
         '[:$-3.1]',
@@ -87,9 +87,9 @@ describe('Unit: util/figure', () => {
           [
             '<figure data-label="fig-1" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
             '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="$-2.0" data-group="$" style="display: none;" data-index="2.0"><div class="figcontent"><div class="math notranslate">$$\nLaTeX\n$$</div></div><span class="figindex">(2.0)</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-b-0.0" data-group="fig" data-index="2.1" id="label:fig-2.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 2.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="$-2.0.0" data-group="$" style="display: none;" data-index="2.0.0"><div class="figcontent"><div class="math notranslate">$$\nLaTeX\n$$</div></div><span class="figindex">(2.0.0)</span><figcaption></figcaption></figure>',
             '<figure data-label="quote-a-0.0" data-group="quote" data-index="2.1" id="label:quote-2.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Quote. 2.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-b-0.0.0" data-group="fig" data-index="2.0.1" id="label:fig-2.0.1"><div class="figcontent"><a href="https://host" rel="noopener" target="_blank"><img class="media" data-src="https://host" alt=""></a></div><span class="figindex">Fig. 2.0.1.</span><figcaption></figcaption></figure>',
             '<figure data-label="$-3.1" data-group="$" data-index="3.1" id="label:$-3.1"><div class="figcontent"><div class="math notranslate">$$\nLaTeX\n$$</div></div><span class="figindex">(3.1)</span><figcaption></figcaption></figure>',
             '<p><a href="#" rel="noopener" class="label" data-label="fig-2.0">fig-2.0</a></p>',
             '<p><a href="#label:$-3.1" rel="noopener" class="label" data-label="$-3.1">(3.1)</a></p>',